+++ kato/kato-gce/kato-gce.gradle
@@ -16,6 +16,7 @@
 
 dependencies {
   compile project(":kato:kato-core")
+  compile project(":clouddriver-google")
   compile spinnaker.dependency("frigga")
 
   spinnaker.group("google")
