+++ kato/kato-gce/src/main/groovy/com/netflix/spinnaker/kato/gce/deploy/GoogleOperationPoller.groovy
@@ -19,11 +19,10 @@ package com.netflix.spinnaker.kato.gce.deploy
 import com.google.api.services.compute.Compute
 import com.google.api.services.compute.model.Operation
 import com.google.api.services.replicapool.Replicapool
+import com.netflix.spinnaker.clouddriver.google.config.GoogleConfigurationProperties
 import com.netflix.spinnaker.kato.data.task.Task
-import com.netflix.spinnaker.kato.gce.deploy.config.GoogleConfig
 import com.netflix.spinnaker.kato.gce.deploy.exception.GoogleOperationException
 import com.netflix.spinnaker.kato.gce.deploy.exception.GoogleOperationTimedOutException
-import com.netflix.spinnaker.kato.gce.deploy.exception.GoogleResourceNotFoundException
 import org.springframework.beans.factory.annotation.Autowired
 
 class GoogleOperationPoller {
@@ -36,7 +35,7 @@ class GoogleOperationPoller {
   }
 
   @Autowired
-  GoogleConfig.GoogleConfigurationProperties googleConfigurationProperties
+  GoogleConfigurationProperties googleConfigurationProperties
 
   private ThreadSleeper threadSleeper = new ThreadSleeper()
