+++ kato/kato-gce/src/test/groovy/com/netflix/spinnaker/kato/gce/deploy/ops/loadbalancer/UpsertGoogleNetworkLoadBalancerAtomicOperationUnitSpec.groovy
@@ -28,10 +28,10 @@ import com.google.api.services.compute.model.Operation
 import com.google.api.services.compute.model.TargetPool
 import com.google.api.services.compute.model.TargetPoolList
 import com.netflix.spinnaker.amos.gce.GoogleCredentials
+import com.netflix.spinnaker.clouddriver.google.config.GoogleConfigurationProperties
 import com.netflix.spinnaker.kato.data.task.Task
 import com.netflix.spinnaker.kato.data.task.TaskRepository
 import com.netflix.spinnaker.kato.gce.deploy.GoogleOperationPoller
-import com.netflix.spinnaker.kato.gce.deploy.config.GoogleConfig
 import com.netflix.spinnaker.kato.gce.deploy.description.UpsertGoogleNetworkLoadBalancerDescription
 import spock.lang.Specification
 import spock.lang.Subject
@@ -110,7 +110,7 @@ class UpsertGoogleNetworkLoadBalancerAtomicOperationUnitSpec extends Specificati
           credentials: credentials)
       @Subject def operation = new UpsertGoogleNetworkLoadBalancerAtomicOperation(description)
       operation.googleOperationPoller =
-        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfig.GoogleConfigurationProperties())
+        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfigurationProperties())
 
     when:
       operation.operate([])
@@ -186,7 +186,7 @@ class UpsertGoogleNetworkLoadBalancerAtomicOperationUnitSpec extends Specificati
           credentials: credentials)
       @Subject def operation = new UpsertGoogleNetworkLoadBalancerAtomicOperation(description)
       operation.googleOperationPoller =
-        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfig.GoogleConfigurationProperties())
+        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfigurationProperties())
 
     when:
       operation.operate([])
@@ -243,7 +243,7 @@ class UpsertGoogleNetworkLoadBalancerAtomicOperationUnitSpec extends Specificati
           credentials: credentials)
       @Subject def operation = new UpsertGoogleNetworkLoadBalancerAtomicOperation(description)
       operation.googleOperationPoller =
-        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfig.GoogleConfigurationProperties())
+        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfigurationProperties())
 
     when:
       operation.operate([])
@@ -290,7 +290,7 @@ class UpsertGoogleNetworkLoadBalancerAtomicOperationUnitSpec extends Specificati
           credentials: credentials)
       @Subject def operation = new UpsertGoogleNetworkLoadBalancerAtomicOperation(description)
       operation.googleOperationPoller =
-        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfig.GoogleConfigurationProperties())
+        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfigurationProperties())
 
     when:
       operation.operate([])
@@ -560,7 +560,7 @@ class UpsertGoogleNetworkLoadBalancerAtomicOperationUnitSpec extends Specificati
         credentials: credentials)
       @Subject def operation = new UpsertGoogleNetworkLoadBalancerAtomicOperation(description)
       operation.googleOperationPoller =
-        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfig.GoogleConfigurationProperties())
+        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfigurationProperties())
 
     when:
       operation.operate([])
@@ -643,7 +643,7 @@ class UpsertGoogleNetworkLoadBalancerAtomicOperationUnitSpec extends Specificati
         credentials: credentials)
       @Subject def operation = new UpsertGoogleNetworkLoadBalancerAtomicOperation(description)
       operation.googleOperationPoller =
-        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfig.GoogleConfigurationProperties())
+        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfigurationProperties())
 
     when:
       operation.operate([])
@@ -733,7 +733,7 @@ class UpsertGoogleNetworkLoadBalancerAtomicOperationUnitSpec extends Specificati
         credentials: credentials)
       @Subject def operation = new UpsertGoogleNetworkLoadBalancerAtomicOperation(description)
       operation.googleOperationPoller =
-        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfig.GoogleConfigurationProperties())
+        new GoogleOperationPoller(googleConfigurationProperties: new GoogleConfigurationProperties())
 
     when:
       operation.operate([])
