+++ oort/oort-gce/src/main/groovy/com/netflix/spinnaker/oort/gce/model/GoogleResourceRetriever.groovy
@@ -27,7 +27,7 @@ import com.google.api.services.replicapool.model.InstanceGroupManagerList
 import com.netflix.frigga.Names
 import com.netflix.spinnaker.amos.AccountCredentialsProvider
 import com.netflix.spinnaker.amos.gce.GoogleCredentials
-import com.netflix.spinnaker.oort.config.GoogleConfig.GoogleConfigurationProperties
+import com.netflix.spinnaker.clouddriver.google.config.GoogleConfigurationProperties
 import com.netflix.spinnaker.oort.gce.model.callbacks.ImagesCallback
 import com.netflix.spinnaker.oort.gce.model.callbacks.InstanceAggregatedListCallback
 import com.netflix.spinnaker.oort.gce.model.callbacks.MIGSCallback
@@ -231,12 +231,12 @@ class GoogleResourceRetriever {
                 it.loadBalancerName == loadBalancerName
               }
 
-              def health = loadBalancerHealth
-                           ? [
+              def health = loadBalancerHealth ?
+                           [
                              state      : loadBalancerHealth.state,
                              description: loadBalancerHealth.description
-                           ]
-                           : [
+                           ] :
+                           [
                              state      : "Unknown",
                              description: "Unable to determine load balancer health."
                            ]
